Refactor screen change, fix splash appearing too late
When there is a change to a different screen (for example when you hit a game over),
only the scene root changes now, not the scene itself. This makes possible the
addition of a fullscreen mode in the future.

The splash was also not appearing as it should on slower pcs, it should be fixed
now I think.