Update Scores record, improve documentation
The Score record isn't serialized for saving anymore, so it no longer implements the
Serializable interface. It also made no sense for the toString override to contain a
space since it's only used for saving, so now the separator is just a semicolon.

Some extra comments were also added to improve the readability of the code.