Create resolution change menu
A new option is now available on the main menu called "RESOLUTION".
When selected, it opens a menu with options for altering the
resolution, or returning to the main menu. If a resolution option
is selected, the window is resized, and so are all elements in
the UI to match the new window size.